--- a/common/src/lazy_pages.rs
+++ b/common/src/lazy_pages.rs
@@ -22,7 +22,7 @@ use crate::Origin;
 use core::fmt;
 use gear_core::{
     ids::ProgramId,
-    memory::{HostPointer, Memory, PageBuf, PageNumber},
+    memory::{HostPointer, Memory, PageBuf, PageNumber, WasmPageNumber},
 };
 use gear_runtime_interface::{gear_ri, RIError};
 use sp_std::{collections::btree_map::BTreeMap, vec::Vec};
@@ -46,11 +46,11 @@ impl From<RIError> for Error {
 }
 
 fn mprotect_lazy_pages(mem: &impl Memory, protect: bool) -> Result<(), Error> {
-    let wasm_mem_addr = match mem.get_buffer_host_addr() {
-        None => return Ok(()),
-        Some(addr) => addr,
-    };
-    gear_ri::mprotect_lazy_pages(wasm_mem_addr, protect)
+    if mem.get_buffer_host_addr().is_none() {
+        return Ok(());
+    }
+
+    gear_ri::mprotect_lazy_pages(protect)
         .map_err(Into::into)
         .map_err(|e| {
             log::error!("{} (it's better to stop node now)", e);
@@ -76,33 +76,25 @@ pub fn is_lazy_pages_enabled() -> bool {
 }
 
 /// Protect and save storage keys for pages which has no data
-pub fn protect_pages_and_init_info(
-    mem: &impl Memory,
-    lazy_pages: impl Iterator<Item = PageNumber>,
-    prog_id: ProgramId,
-) -> Result<(), Error> {
-    let prog_id_hash = prog_id.into_origin();
-    let mut lay_pages_peekable = lazy_pages.peekable();
-
+pub fn protect_pages_and_init_info(mem: &impl Memory, prog_id: ProgramId) -> Result<(), Error> {
     gear_ri::reset_lazy_pages_info();
 
-    let addr = match mem.get_buffer_host_addr() {
-        None => {
-            return if lay_pages_peekable.peek().is_none() {
-                // In this case wasm buffer cannot be undefined
-                Err(Error::WasmMemBufferIsUndefined)
-            } else {
-                Ok(())
-            };
-        }
-        Some(addr) => addr,
+    let prog_prefix = crate::pages_prefix(prog_id.into_origin());
+    gear_ri::set_program_prefix(prog_prefix);
+
+    let addr = if let Some(addr) = mem.get_buffer_host_addr() {
+        addr
+    } else {
+        return Ok(());
     };
+
     gear_ri::set_wasm_mem_begin_addr(addr).map_err(|e| {
         log::error!("{} (it's better to stop node now)", e);
         e
     })?;
 
-    crate::save_page_lazy_info(prog_id_hash, lay_pages_peekable);
+    let size = (mem.size().0 + 1) * WasmPageNumber::size() as u32;
+    gear_ri::set_wasm_mem_size(size)?;
 
     mprotect_lazy_pages(mem, true)
 }
@@ -131,10 +123,11 @@ pub fn remove_lazy_pages_prot(mem: &impl Memory) -> Result<(), Error> {
     mprotect_lazy_pages(mem, false)
 }
 
-/// Protect lazy-pages and set new wasm mem addr if it has been changed
-pub fn protect_lazy_pages_and_update_wasm_mem_addr(
+/// Protect lazy-pages and set new wasm mem addr and size, if they have been changed
+pub fn update_lazy_pages_and_protect_again(
     mem: &impl Memory,
     old_mem_addr: Option<HostPointer>,
+    old_mem_size: WasmPageNumber,
 ) -> Result<(), Error> {
     struct PointerDisplay(HostPointer);
 
@@ -157,15 +150,14 @@ pub fn protect_lazy_pages_and_update_wasm_mem_addr(
                 e
             })?;
     }
-    mprotect_lazy_pages(mem, true)
-}
 
-/// Returns list of current lazy pages numbers
-pub fn get_lazy_pages_numbers() -> Vec<PageNumber> {
-    gear_ri::get_lazy_pages_numbers()
-        .iter()
-        .map(|p| PageNumber(*p))
-        .collect()
+    let new_mem_size = mem.size();
+    if new_mem_size > old_mem_size {
+        let size = (new_mem_size.0 + 1) * WasmPageNumber::size() as u32;
+        gear_ri::set_wasm_mem_size(size)?;
+    }
+
+    mprotect_lazy_pages(mem, true)
 }
 
 /// Returns list of realeased pages numbers