--- a/core-processor/src/common.rs
+++ b/core-processor/src/common.rs
@@ -338,7 +338,7 @@ pub enum ExecutionErrorReason {
     OutOfRent,
     /// Initial pages data must be empty when execute with lazy pages
     #[display(fmt = "Initial pages data must be empty when execute with lazy pages")]
-    InitialPagesDataWhenLazyPages,
+    InitialPagesContainsDataInLazyPagesMode,
 }
 
 /// Actor.