@ukint-vs 2022-07-06
the fact that the number of samples has decreased does not mean that the performance has grown, `perf` cannot sample signal handler. Therefore it is impossible to measure what is happening inside the handler.