@ukint-vs 2022-07-06
> > 
> 
> What do you mean `side by side` ?

this branch:  `+ r    58.08 µs`
master: `+ r     57.9 µs`